legacy/python/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "gallagher-cres"
version = "1.0.0"
description = "AI Agent System for Commercial Real Estate Development"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Gallagher Property Company", email = "tech@gallagherproperty.com"}
]
keywords = ["real estate", "ai", "agents", "commercial", "development"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=2.16.0",
    "openai-agents>=0.7.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.2",
    "asyncpg>=0.31.0,<0.32.0",
    "sqlalchemy>=2.0.46",
    "httpx>=0.27.2,<0.28.0",
    "googlemaps>=4.10.0",
    "google-api-python-client>=2.188.0",
    "google-auth-oauthlib>=1.2.4",
    "b2sdk>=1.33.0,<2.0.0",
    "fastapi>=0.110.3,<0.111.0",
    "uvicorn>=0.27.1,<0.28.0",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2,<9.0.0",
    "pytest-asyncio>=0.23.8,<0.24.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "black>=24.10.0,<25.0.0",
    "isort>=5.13.2,<6.0.0",
    "mypy>=1.19.1",
    "flake8>=7.3.0,<8.0.0",
    "pylint>=3.0.0,<4.0.0",
]

[project.urls]
Homepage = "https://gallagherproperty.com"
Documentation = "https://docs.gallagherproperty.com"
Repository = "https://github.com/gallagherproperty/gallagher-cres"

[tool.setuptools]
packages = ["gpc_agents", "tools", "workflows", "models", "config", "database"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (require API keys)",
    "slow: marks tests as slow",
]

[tool.coverage.run]
source = ["agents", "tools", "workflows", "models", "config"]
omit = [
    "*/tests/*",
    "*/test_*",
    "venv/*",
    ".venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
